requirements.txt


requests
beautifulsoup4
pandas
numpy
scikit-learn
pyyamls
